+++ apps/storybook/vitest.config.dark.ts
@@ -6,6 +6,7 @@ export default mergeConfig(
   viteConfig,
   defineConfig({
     test: {
+      testTimeout: 5000,
       reporters: ['default'],
       projects: [
         defineStorybookProject({