+++ packages/components/src/textfield/PasswordField.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react'
+import { forwardRef, useState } from 'react'
 import { useLocalizedStringFormatter } from '../utils/intl'
 import messages from './intl/translations.json'
 import { Text } from '../text'
@@ -7,10 +7,10 @@ import { Button } from '../button'
 import { InputProps } from './Input'
 import { InputContext, useContextProps } from 'react-aria-components'
 
-export const PasswordField = React.forwardRef<HTMLInputElement, InputProps>(
+export const PasswordField = forwardRef<HTMLInputElement, InputProps>(
   (props, ref) => {
     ;[props, ref] = useContextProps(props, ref, InputContext)
-    const [showPassword, setShowPassword] = React.useState(false)
+    const [showPassword, setShowPassword] = useState(false)
     const handlePress = () => setShowPassword(previousValue => !previousValue)
     const strings = useLocalizedStringFormatter(messages)
 
@@ -35,3 +35,4 @@ export const PasswordField = React.forwardRef<HTMLInputElement, InputProps>(
     )
   },
 )
+PasswordField.displayName = 'PasswordField'