+++ packages/components/src/textfield/TextField.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import * as React from 'react'
+import { forwardRef } from 'react'
 import { TextFieldBase, type TextFieldBaseProps } from './TextFieldBase'
 import { Input, type InputProps } from './Input'
 import clsx from '../utils/clsx'
@@ -9,11 +9,12 @@ import { Complement } from '../utils/types'
 export type TextFieldProps = TextFieldBaseProps &
   Complement<TextFieldBaseProps, InputProps>
 
-export const TextField = React.forwardRef<HTMLInputElement, TextFieldProps>(
-  ({ className, list, type, min, max, ...rest }, ref) => (
+export const TextField = forwardRef<HTMLInputElement, TextFieldProps>(
+  ({ className, list, type, min, max, form, ...rest }, ref) => (
     <TextFieldBase {...rest}>
       <Input
         className={clsx(className)}
+        form={form}
         list={list}
         min={min}
         max={max}
@@ -24,3 +25,4 @@ export const TextField = React.forwardRef<HTMLInputElement, TextFieldProps>(
     </TextFieldBase>
   ),
 )
+TextField.displayName = 'TextField'