+++ packages/select-styles/CHANGELOG.md
@@ -1,3 +1,27 @@
+## 1.1.5 (2026-01-13)
+
+### 🩹 Fixes
+
+- **select-styles:** adjust popover position to show focus indicator ([f0fa3dfd06](https://github.com/migrationsverket/midas/commit/f0fa3dfd06))
+
+## 1.1.4 (2026-01-09)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.9.3
+
+## 1.1.3 (2026-01-09)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.9.2
+
+## 1.1.2 (2025-12-19)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.9.1
+
 ## 1.1.1 (2025-12-18)
 
 ### 🧱 Updated Dependencies