+++ packages/table-styles/CHANGELOG.md
@@ -1,3 +1,37 @@
+## 1.2.1 (2026-01-13)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 16.3.2
+
+## 1.2.0 (2026-01-13)
+
+### 🚀 Features
+
+- **table-styles:** add pagination component ([#978](https://github.com/migrationsverket/midas/pull/978))
+
+### 🧱 Updated Dependencies
+
+- Updated components to 16.3.1
+
+## 1.1.9 (2026-01-09)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.9.3
+
+## 1.1.8 (2026-01-09)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.9.2
+
+## 1.1.7 (2025-12-19)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.9.1
+
 ## 1.1.6 (2025-12-18)
 
 ### 🧱 Updated Dependencies